requirements.txt
Django==1.10.5
django-braces==1.10.0
django-compat==1.0.13
django-jquery==3.1.0
django-front==0.5.3
django-hijack==2.1.2
django-widget-tweaks==1.4.1
django-constance[database]==1.3.4
whitenoise==3.2.3
Pillow==4.0.0
mysqlclient==1.3.9
django-autocomplete-light==3.2.1
ldap3==2.2.0
